requirements.txt
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
django-model-utils==4.0.0